Pull code down into entity
